Day 7 parts 1 & 2

+++ Day.h
@@ -15,6 +15,6 @@ public:
     [[nodiscard]] virtual auto num() const -> int = 0;
     virtual auto example_input_str1() const -> std::stringstream = 0;
     virtual auto example_input_str2() const -> std::stringstream {return example_input_str1();}
-    virtual auto part1(const Input& input) const -> int = 0;
-    virtual auto part2(const Input& input) const -> int = 0;
+    virtual auto part1(const Input& input) const -> long long = 0;
+    virtual auto part2(const Input& input) const -> long long = 0;
 };
